Move old RulesValidator, implement new interface.
